+++ NEWS.md
@@ -20,6 +20,10 @@
 - Add `predict` method to obtain conditional mean estimates and prediction intervals, both with or without previous observations.
 - Use automatic differentiation to calculate Satterthwaite adjusted degrees of freedom calculation, resulting in 10-fold speed-up of the Satterthwaite calculations after the initial model fit.
 
+### Miscellaneous
+
+- Removed `free_cores()` in favor of `parallelly::availableCores(omit = 1)`.
+
 ### Bug Fixes
 
 - Previously `mmrm` fit follows the global option `na.action` and if it is set